+++ core/src/main/java/uk/ac/bbsrc/tgac/miso/core/data/QcCorrespondingField.java
@@ -1,10 +1,12 @@
 package uk.ac.bbsrc.tgac.miso.core.data;
 
+import java.math.BigDecimal;
+
 public enum QcCorrespondingField {
   CONCENTRATION {
 
     @Override
-    public void updateField(Pool pool, double value, String units) {
+    public void updateField(Pool pool, BigDecimal value, String units) {
       pool.setConcentration(value);
       ConcentrationUnit concUnit = ConcentrationUnit.getFromString(units);
       if (concUnit != null) {
@@ -13,7 +15,7 @@ public enum QcCorrespondingField {
     }
 
     @Override
-    public void updateField(Library library, double value, String units) {
+    public void updateField(Library library, BigDecimal value, String units) {
       library.setConcentration(value);
       ConcentrationUnit concUnit = ConcentrationUnit.getFromString(units);
       if (concUnit != null) {
@@ -22,7 +24,7 @@ public enum QcCorrespondingField {
     }
 
     @Override
-    public void updateField(Sample sample, double value, String units) {
+    public void updateField(Sample sample, BigDecimal value, String units) {
       sample.setConcentration(value);
       ConcentrationUnit concUnit = ConcentrationUnit.getFromString(units);
       if (concUnit != null) {
@@ -34,7 +36,7 @@ public enum QcCorrespondingField {
   VOLUME {
 
     @Override
-    public void updateField(Pool pool, double value, String units) {
+    public void updateField(Pool pool, BigDecimal value, String units) {
       pool.setVolume(value);
       VolumeUnit volUnit = VolumeUnit.getFromString(units);
       if (volUnit != null) {
@@ -43,7 +45,7 @@ public enum QcCorrespondingField {
     }
 
     @Override
-    public void updateField(Library library, double value, String units) {
+    public void updateField(Library library, BigDecimal value, String units) {
       library.setVolume(value);
       VolumeUnit volUnit = VolumeUnit.getFromString(units);
       if (volUnit != null) {
@@ -52,7 +54,7 @@ public enum QcCorrespondingField {
     }
 
     @Override
-    public void updateField(Sample sample, double value, String units) {
+    public void updateField(Sample sample, BigDecimal value, String units) {
       sample.setVolume(value);
       VolumeUnit volUnit = VolumeUnit.getFromString(units);
       if (volUnit != null) {
@@ -63,23 +65,23 @@ public enum QcCorrespondingField {
   },
   NONE;
 
-  public void updateField(Pool pool, double value, String units) {
+  public void updateField(Pool pool, BigDecimal value, String units) {
     throw new UnsupportedOperationException("Method not implemented for unspecified field");
   }
 
-  public void updateField(Library library, double value, String units) {
+  public void updateField(Library library, BigDecimal value, String units) {
     throw new UnsupportedOperationException("Method not implemented for unspecified field");
   }
 
-  public void updateField(Sample sample, double value, String units) {
+  public void updateField(Sample sample, BigDecimal value, String units) {
     throw new UnsupportedOperationException("Method not implemented for unspecified field");
   }
 
-  public void updateField(Run run, double value, String units) {
+  public void updateField(Run run, BigDecimal value, String units) {
     throw new UnsupportedOperationException("Method not implemented for unspecified field");
   }
 
-  public void updateField(SequencerPartitionContainer container, double value, String units) {
+  public void updateField(SequencerPartitionContainer container, BigDecimal value, String units) {
     throw new UnsupportedOperationException("Method not implemented for unspecified field");
   }
