+++ sqlstore/src/main/java/uk/ac/bbsrc/tgac/miso/persistence/impl/HibernateLibraryQcDao.java
@@ -1,6 +1,7 @@
 package uk.ac.bbsrc.tgac.miso.persistence.impl;
 
 import java.io.IOException;
+import java.math.BigDecimal;
 import java.util.Collection;
 
 import org.hibernate.Session;
@@ -73,7 +74,7 @@ public class HibernateLibraryQcDao implements LibraryQcStore {
   }
 
   @Override
-  public void updateEntity(long id, QcCorrespondingField correspondingField, double value, String units) throws IOException {
+  public void updateEntity(long id, QcCorrespondingField correspondingField, BigDecimal value, String units) throws IOException {
     Library library = getLibrary(id);
     correspondingField.updateField(library, value, units);
     currentSession().update(library);
